Exception handled: Delete

+++ src/main/java/com/vhbrito/workshop/services/UserService.java
@@ -2,8 +2,11 @@ package com.vhbrito.workshop.services;
 
 import com.vhbrito.workshop.entities.User;
 import com.vhbrito.workshop.repositories.UserRepository;
+import com.vhbrito.workshop.services.exceptions.DatabaseException;
 import com.vhbrito.workshop.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,8 +32,15 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
-    }
+        try {
+            if (userRepository.existsById(id)) userRepository.deleteById(id);
+            else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
+}
 
     public User updateUser(Long id, User user) {
         User entity = userRepository.getReferenceById(id);
